Refactor authentication flow: rename `loginAdmin` to `loginResult`, extract logout logic to `LogoutController`, remove unused dependencies, and improve code readability across AuthController and LoginController.
