+++ src/main/java/local/code_compass_backend/controller/AuthController.java
@@ -1,14 +1,11 @@
 package local.code_compass_backend.controller;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import local.code_compass_backend.client.SBAuthClient;
 import local.code_compass_backend.dto.AuthDto;
-import local.code_compass_backend.dto.ProfileDto;
 import local.code_compass_backend.service.AuthService;
 import local.code_compass_backend.utility.CookieUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -29,8 +26,8 @@ public class AuthController {
     private record LoginResponse(User user) {}
 
     @PostMapping("/api/login")
-    public ResponseEntity<?> loginAdmin(@RequestBody AuthDto authDto) {
-        AuthService.LoginResult result = authService.loginAdmin(authDto);
+    public ResponseEntity<?> loginResult(@RequestBody AuthDto authDto) {
+        AuthService.LoginResult result = authService.loginResult(authDto);
         LoginResponse body = new LoginResponse(new User(result.email(), result.displayName()));
         return cookieUtil
                 .setJwtCookie(ResponseEntity.ok(), result.accessToken())
