+++ src/main/java/local/code_compass_backend/controller/LoginController.java
@@ -1,13 +1,9 @@
 package local.code_compass_backend.controller;
 
-
 import local.code_compass_backend.dto.ProfileDto;
-import local.code_compass_backend.utility.CookieUtil;
 import local.code_compass_backend.service.LoginService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -18,29 +14,9 @@ public class LoginController {
  @Autowired
     private LoginService loginService;
 
-    @Autowired
-    private CookieUtil cookieUtil;
-
-    @PostMapping ("/api/trainee_login")
+ @PostMapping ("/api/trainee_login")
     public ResponseEntity<Void>  validateLogIn(@RequestBody ProfileDto profileDto) {
         loginService.validateLogIn(profileDto);
         return ResponseEntity.ok().build();
     }
-
-    @PostMapping("/api/logout")
-    public ResponseEntity<?> logout() {
-        ResponseCookie cookie = cookieUtil.clearJwtCookie();
-
-        return ResponseEntity.ok()
-                .header("Set-Cookie", cookie.toString())
-                .body("Logged out");
-    }
-
-
-
-
-
-
-    //gebruiker aanmaken = iemand registreert zich, dan moet de backend een bericht sturen naar SupaBase met "registreer deze gebruiker"
-    //de UUID die dan terugkomt moet de UserId worden, dit is een 1 op 1 relatie (Dat is dus die PK/FK relatie)
 }
