+++ src/main/java/local/code_compass_backend/controller/LogoutController.java
@@ -1,0 +1,24 @@
+package local.code_compass_backend.controller;
+import local.code_compass_backend.utility.CookieUtil;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseCookie;
+
+import org.springframework.http.ResponseCookie;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class LogoutController {
+    @Autowired
+    private CookieUtil cookieUtil;
+
+    @PostMapping("/api/logout")
+    public ResponseEntity<?> logout() {
+        ResponseCookie cookie = cookieUtil.clearJwtCookie();
+
+        return ResponseEntity.ok()
+                .header("Set-Cookie", cookie.toString())
+                .body("Logged out");
+    }
+}
