+++ src/main/java/local/code_compass_backend/controller/NewUserController.java
@@ -1,0 +1,8 @@
+package local.code_compass_backend.controller;
+
+public class NewUserController {
+}
+
+
+//gebruiker aanmaken = iemand registreert zich, dan moet de backend een bericht sturen naar SupaBase met "registreer deze gebruiker"
+//de UUID die dan terugkomt moet de UserId worden, dit is een 1 op 1 relatie (Dat is dus die PK/FK relatie)
