+++ src/main/java/local/code_compass_backend/service/AuthService.java
@@ -20,11 +20,11 @@ public class AuthService {
     public record LoginResult(String accessToken, String email, String displayName) {}
 
 
-    public LoginResult loginAdmin(AuthDto authDto) {
-        SBAuthClient.AuthResult auth = authenticateAdmin(authDto);
-        var profile = profileRepository.findById(auth.getUserId())
+    public LoginResult loginResult(AuthDto authDto) {
+        SBAuthClient.AuthResult userLogin = authenticateAdmin(authDto);
+        var profile = profileRepository.findById(userLogin.getUserId())
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Profiel niet gevonden"));
-        return new LoginResult(auth.getAccessToken(), profile.getEmail(), profile.getDisplayName());
+        return new LoginResult(userLogin.getAccessToken(), profile.getEmail(), profile.getDisplayName());
     }
     public SBAuthClient.AuthResult authenticateAdmin(AuthDto authDto) {
         if (authDto == null || authDto.getEmail() == null || authDto.getEmail().isBlank()
@@ -32,18 +32,19 @@ public class AuthService {
             throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Voer je gegevens in.");
         }
 
-        SBAuthClient.AuthResult auth;
+        SBAuthClient.AuthResult authenticationResult;
         try {
-            auth = sbAuthClient.authenticateAndGetUser(authDto.getEmail(), authDto.getPassword());
+            authenticationResult = sbAuthClient.authenticateAndGetUser(authDto.getEmail(), authDto.getPassword());
         } catch (HttpStatusCodeException | IllegalStateException e) {
             throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Ongeldige inloggegevens.");
         }
 
-        boolean isAdmin = profileRepository.existsByIdAndRole(auth.getUserId(), Role.ADMIN);
+
+        boolean isAdmin = profileRepository.existsByIdAndRole(authenticationResult.getUserId(), Role.ADMIN);
         if (!isAdmin) {
             throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Onvoldoende rechten.");
         }
-        return auth;
+        return authenticationResult;
     }
 
 }
